update Footer.tsx and pages/index.tsx
